@@ -1,23 +1,21 @@
-'use strict';
-
-const path = require('path');
-const AutoLoad = require('@fastify/autoload');
-
-/**
- * @typedef {import('fastify').FastifyInstance} FastifyInstance
- */
+import path from 'path';
+import AutoLoad from '@fastify/autoload';
+import FastifyEnv from '@fastify/env';
+import { fileURLToPath } from 'url';
+const __dirname = path.dirname(fileURLToPath(import.meta.url));
 
 // Pass --options via CLI arguments in command to enable these options.
-module.exports.options = {};
+export const options = {};
 
 /**
- * @param {FastifyInstance} fastify
+ * @param {import('fastify').FastifyInstance} fastify
  * @param {Object} opts
  */
-module.exports = async function (fastify, opts) {
-  // Place here your custom code!
+export default async function (fastify, opts) {
+  /** @type {import('@fastify/env').FastifyEnvOptions} */
   const envOptions = {
-    confKey: 'config', // optional, default: 'config'
+    confKey: 'config',
+    data: process.env,
     schema: {
       type: 'object',
       required: ['PORT', 'MONGODB_URL', 'JWT_SECRET', 'GOOGLE_CLIENT_ID', 'GOOGLE_CLIENT_SECRET'],
@@ -42,7 +40,7 @@ module.exports = async function (fastify, opts) {
           type: 'string',
           default: '155453cr37'
         },
-      }
+      },
     },
     // will read .env in root folder
     dotenv: {
@@ -50,22 +48,18 @@ module.exports = async function (fastify, opts) {
       debug: true
     }
   };
-  await fastify.register(require('@fastify/env'), envOptions);
+  await fastify.register(FastifyEnv, envOptions);
 
-  // Do not touch the following lines
-
-  // This loads all plugins defined in plugins
-  // those should be support plugins that are reused
-  // through your application
+  // ~~~ Do not touch the following lines ~~~ //
+  // loads all plugins defined in plugins
   fastify.register(AutoLoad, {
     dir: path.join(__dirname, 'plugins'),
     options: Object.assign({}, opts)
   });
 
-  // This loads all plugins defined in routes
-  // define your routes in one of these
+  // loads all plugins defined in routes
   fastify.register(AutoLoad, {
     dir: path.join(__dirname, 'routes'),
     options: Object.assign({}, opts)
   });
-};
+}
