@@ -1,5 +1,6 @@
 /** @type {import("mongoose").Mongoose} */
-const mongoose = require('mongoose');
+import mongoose from 'mongoose';
+
 /**
  * Mongoose schema for user accounts.
  *
@@ -31,7 +32,6 @@ const userSchema = new mongoose.Schema({
   },
   profilePhoto: {
     type: String,
-    unique: true,
   },
   uploads: [{
     type: mongoose.Schema.Types.ObjectId,
@@ -43,4 +43,4 @@ const userSchema = new mongoose.Schema({
   }],
 });
 
-module.exports = mongoose.model('users', userSchema);
+export default mongoose.model('users', userSchema);
