@@ -1,9 +1,7 @@
-'use strict';
+import fp from 'fastify-plugin';
+import mongoose from 'mongoose';
 
-const fp = require('fastify-plugin');
-const mongoose = require('mongoose');
-
-module.exports = fp(async function (fastify, options) {
+export default fp(async function (fastify, options) {
   // Log a message when the MongoDB connection is established
   mongoose.connection.on('connected', () => {
     fastify.log.info('MongoDB connection established!');
@@ -13,26 +11,6 @@ module.exports = fp(async function (fastify, options) {
     const imageBucket = new mongoose.mongo.GridFSBucket(mongoose.connection.db, {
       bucketName: 'images'
     });
-    // uncomment to see all current files in the bucket
-    // soundBucket.find({}).forEach((file) => {
-    //   fastify.log.info(file);
-    // });
-    // imageBucket.find({}).forEach((file) => {
-    //   fastify.log.info(file);
-    // });
-    // uncomment to drop the bucket
-    // soundBucket.drop((err, result) => {
-    //   if (err) {
-    //     fastify.log.error(err);
-    //   }
-    //   fastify.log.info(result);
-    // });
-    // imageBucket.drop((err, result) => {
-    //   if (err) {
-    //     fastify.log.error(err);
-    //   }
-    //   fastify.log.info(result);
-    // });
     fastify.decorate('gridfsSounds', soundBucket);
     fastify.decorate('gridfsImages', imageBucket);
     fastify.decorate('toObjectId', (val) => new mongoose.Types.ObjectId(val));
