@@ -1,11 +1,9 @@
-'use strict';
-
-const fp = require('fastify-plugin');
+import fp from 'fastify-plugin';
 
 // the use of fastify-plugin is required to be able
 // to export the decorators to the outer scope
 
-module.exports = fp(async function (fastify, opts) {
+export default fp(async function (fastify, opts) {
   fastify.decorate('someSupport', function () {
     return 'hugs';
   });
