@@ -1,6 +1,10 @@
-require('dotenv').config({ path: './local.env' });
-const mongoose = require('mongoose');
+import pkg from 'dotenv';
+import mongoose from 'mongoose';
+import User from '../models/User.js';
+import Sound from '../models/Sound.js';
+import Image from '../models/Image.js';
 
+pkg.config({ path: './local.env' });
 /**
  * Cleans the MongoDB database.
  * @async
@@ -9,10 +13,7 @@ async function cleanDatabase () {
   try {
     await mongoose.connect(process.env.MONGODB_URL, { useNewUrlParser: true, useUnifiedTopology: true });
     console.log('Connected to MongoDB');
-    const User = require('../models/User');
-    const Sound = require('../models/Sound');
-    const Image = require('../models/Image');
-
+    
     const soundBucket = new mongoose.mongo.GridFSBucket(mongoose.connection.db, {
       bucketName: 'sounds'
     });
