@@ -1,8 +1,6 @@
-'use strict';
-
-const { test } = require('tap');
-const Fastify = require('fastify');
-const Support = require('../../plugins/support');
+import { test } from 'tap';
+import Fastify from 'fastify';
+import Support from '../../plugins/support.js';
 
 test('support works standalone', async (t) => {
   const fastify = Fastify();
@@ -11,16 +9,3 @@ test('support works standalone', async (t) => {
   await fastify.ready();
   t.equal(fastify.someSupport(), 'hugs');
 });
-
-// You can also use plugin with opts in fastify v2
-//
-// test('support works standalone', (t) => {
-//   t.plan(2)
-//   const fastify = Fastify()
-//   fastify.register(Support)
-//
-//   fastify.ready((err) => {
-//     t.error(err)
-//     t.equal(fastify.someSupport(), 'hugs')
-//   })
-// })
