@@ -1,7 +1,5 @@
-'use strict';
-
-const { test } = require('tap');
-const { build } = require('../helper');
+import { test } from 'tap';
+import { build } from '../helper.js';
 
 test('example is loaded', async (t) => {
   const app = await build(t);
@@ -11,17 +9,3 @@ test('example is loaded', async (t) => {
   });
   t.equal(res.payload, 'this is an example');
 });
-
-// inject callback style:
-//
-// test('example is loaded', (t) => {
-//   t.plan(2)
-//   const app = await build(t)
-//
-//   app.inject({
-//     url: '/example'
-//   }, (err, res) => {
-//     t.error(err)
-//     t.equal(res.payload, 'this is an example')
-//   })
-// })
