@@ -1,7 +1,5 @@
-'use strict';
-
-const { test } = require('tap');
-const { build } = require('../helper');
+import { test } from 'tap';
+import { build } from '../helper.js';
 
 test('default root route', async (t) => {
   const app = await build(t);
@@ -9,19 +7,5 @@ test('default root route', async (t) => {
   const res = await app.inject({
     url: '/'
   });
-  t.same(JSON.parse(res.payload), { root: true });
+  t.same(JSON.parse(res.payload), { root: 'Hello World' });
 });
-
-// inject callback style:
-//
-// test('default root route', (t) => {
-//   t.plan(2)
-//   const app = await build(t)
-//
-//   app.inject({
-//     url: '/'
-//   }, (err, res) => {
-//     t.error(err)
-//     t.same(JSON.parse(res.payload), { root: true })
-//   })
-// })
