@@ -1,7 +1,7 @@
 <template>
-  <v-app fullscreen>
-    <v-main style="height: 100vh; width: 100vw; overflow-y: hidden;">
-      <v-container fluid style="padding: 0;">
+  <v-app fullscreen style="width:100vw">
+    <v-main>
+      <v-container fluid style="padding: 0; height: 100%">
         <router-view></router-view>
       </v-container>
     </v-main>
@@ -33,7 +33,14 @@ export default {
   text-align: center;
   color: #2c3e50;
 }
+
 .v-container {
   background-color: cyan;
+}
+</style>
+<style scoped>
+.v-main{
+  display: flex;
+  flex-direction: column;
 }
 </style>
