@@ -1,11 +1,11 @@
 <template>
-  <footer>
+  <v-footer app>
     <p>
       <span @click="nav()">
         &copy; 2023 Sound Map for a Changing Landscape
       </span>
     </p>
-  </footer>
+  </v-footer>
 </template>
 
 <script>
@@ -25,14 +25,13 @@ export default {
 
 <!-- Add "scoped" attribute to limit CSS to this component only -->
 <style scoped>
-footer {
-  position: fixed;
-  bottom: 0;
+.v-footer {
   width: 100%;
   height: 2.5rem;
   line-height: 2.5rem;
   background-color: #00bcd4;
   text-align: center;
+  justify-content: center;
 }
 span {
   cursor: pointer;
